rename packages at konsist-path-tester
