+++ test-projects/konsist-path-tester/app/src/integrationTest/kotlin/com/lemonappdev/konsist/container/from/KoScopeFromModuleTest.kt
@@ -31,10 +31,10 @@ class KoScopeFromModuleTest {
                 "$appIntegrationTestSourceSetDirectory/konsist/container/from/KoScopeFromProjectTest.kt",
                 "$appIntegrationTestSourceSetDirectory/konsist/container/from/KoScopeFromSourceSetTest.kt",
                 "$appIntegrationTestSourceSetDirectory/konsist/container/from/KoScopeFromTest.kt",
-                "$appIntegrationTestSourceSetDirectory/konsist/declaration/koclassdeclaration/" +
+                "$appIntegrationTestSourceSetDirectory/konsist/declaration/koclass/" +
                         "KoClassDeclarationForKoHasTestClassProviderTest.kt",
-                "$appIntegrationTestSourceSetDirectory/konsist/declaration/kofiledeclaration/KoFileForKoModuleProviderTest.kt",
-                "$appIntegrationTestSourceSetDirectory/konsist/declaration/kofiledeclaration/KoFileForKoSourceSetProviderTest.kt",
+                "$appIntegrationTestSourceSetDirectory/konsist/declaration/kofile/KoFileForKoModuleProviderTest.kt",
+                "$appIntegrationTestSourceSetDirectory/konsist/declaration/kofile/KoFileForKoSourceSetProviderTest.kt",
                 "$appIntegrationTestSourceSetDirectory/konsist/helper/ext/KoScopeExt.kt",
                 "$appIntegrationTestSourceSetDirectory/konsist/helper/ext/PathExt.kt",
                 "$appIntegrationTestSourceSetDirectory/konsist/helper/util/PathProvider.kt",
@@ -101,10 +101,10 @@ class KoScopeFromModuleTest {
                 "$appIntegrationTestSourceSetDirectory/konsist/container/from/KoScopeFromProjectTest.kt",
                 "$appIntegrationTestSourceSetDirectory/konsist/container/from/KoScopeFromSourceSetTest.kt",
                 "$appIntegrationTestSourceSetDirectory/konsist/container/from/KoScopeFromTest.kt",
-                "$appIntegrationTestSourceSetDirectory/konsist/declaration/koclassdeclaration/" +
+                "$appIntegrationTestSourceSetDirectory/konsist/declaration/koclass/" +
                         "KoClassDeclarationForKoHasTestClassProviderTest.kt",
-                "$appIntegrationTestSourceSetDirectory/konsist/declaration/kofiledeclaration/KoFileForKoModuleProviderTest.kt",
-                "$appIntegrationTestSourceSetDirectory/konsist/declaration/kofiledeclaration/KoFileForKoSourceSetProviderTest.kt",
+                "$appIntegrationTestSourceSetDirectory/konsist/declaration/kofile/KoFileForKoModuleProviderTest.kt",
+                "$appIntegrationTestSourceSetDirectory/konsist/declaration/kofile/KoFileForKoSourceSetProviderTest.kt",
                 "$appIntegrationTestSourceSetDirectory/konsist/helper/ext/KoScopeExt.kt",
                 "$appIntegrationTestSourceSetDirectory/konsist/helper/ext/PathExt.kt",
                 "$appIntegrationTestSourceSetDirectory/konsist/helper/util/PathProvider.kt",
